fix: declare support for 0.81
