+++ example/index.ts
@@ -1,6 +1,6 @@
 // @react-native-webapis
 import { AppRegistry } from "react-native";
-import { App } from "./App";
 import { name as appName } from "./app.json";
+import { App } from "./src/App";
 
 AppRegistry.registerComponent(appName, () => App);
