+++ example/src/Separator.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { View } from "react-native";
+import { useStyles } from "./styles";
+
+export function Separator(): React.ReactElement {
+  const styles = useStyles();
+  return <View style={styles.separator} />;
+}
