+++ package.json
@@ -101,7 +101,7 @@
     "@callstack/react-native-visionos": "0.73 - 0.78",
     "@expo/config-plugins": ">=5.0",
     "react": "18.1 - 19.1",
-    "react-native": "0.70 - 0.80 || >=0.81.0-0 <0.81.0",
+    "react-native": "0.70 - 0.81 || >=0.82.0-0 <0.82.0",
     "react-native-macos": "^0.0.0-0 || 0.71 - 0.78",
     "react-native-windows": "^0.0.0-0 || 0.70 - 0.79"
   },
